# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.SessionListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/dto.SessionResponse'
        type: array
    type: object
  dto.SessionResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_current:
        type: boolean
      last_active:
        type: string
      user_agent:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample authentication and authorization API built with Go
    and Gin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication and Authorization API
  version: 1.1.0
paths:
...
  /sessions:
    delete:
      description: Revoke all sessions except the current one
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Revoke all other sessions
      tags:
      - Sessions
    get:
      description: List all active sessions (devices/IPs) for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SessionListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List active sessions
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      description: Revoke a specific session by its ID (logout a specific device)
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Revoke a specific session
      tags:
      - Sessions
...
securityDefinitions:
  AdminApiKey:
    description: Admin API Key for protected admin routes
    in: header
    name: X-Admin-API-Key
    type: apiKey
  ApiKeyAuth:
    description: Type "Bearer" + your JWT token
    in: header
    name: Authorization
    type: apiKey
  AppApiKey:
    description: Per-application API Key for app-scoped routes
    in: header
    name: X-App-API-Key
    type: apiKey
  AppID:
    description: The UUID of the application context
    in: header
    name: X-App-ID
    type: apiKey
swagger: "2.0"